requirements.txt
tqdm==4.31.1
matplotlib==3.3.1
scipy==1.5.2
onnxruntime==1.4.0
numpy==1.19.1
torchvision==0.8.0a0
opencv_python==3.4.1.15
Pillow==8.3.1
PyYAML==5.4.1
scikit_learn==0.24.2
tensorboardX==2.4
openvino==2021.4.0
